@@ -4,30 +4,23 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace HahnMovies.Infrastructure.Data.Config;
 
-public class MoviesConfig : IEntityTypeConfiguration<Movies>
+public class MoviesConfig : IEntityTypeConfiguration<Movie>
 {
-    public void Configure(EntityTypeBuilder<Movies> builder)
-    {
-        // Configure table name (optional)
-        builder.ToTable("Movies");
+  
+        public void Configure(EntityTypeBuilder<Movie> builder)
+        {
+            
+            builder.HasKey(m => m.Id);
+            
+            builder.Property(m => m.Id)
+                .ValueGeneratedNever();
+            
+            builder.Property(m => m.Title)
+                .IsRequired(false)
+                .HasMaxLength(200);
 
-        // Configure properties
-
-        builder.HasKey(m => m.Id);
-        builder.Property(m => m.Title)
-            .HasMaxLength(200)
-            .IsRequired();
-
-        builder.Property(m => m.ReleaseDate)
-            .IsRequired();
-
-        builder.Property(m => m.VoteAverage)
-            .HasPrecision(5, 2); // Example: Precision for decimal values
-
-        builder.Property(m => m.PosterPath)
-            .HasMaxLength(500)
-            .IsRequired();
-
-        // Add any additional configurations here
+            builder.Property(m => m.PosterPath)
+                .HasMaxLength(100)
+                .IsRequired(false);
+        }
     }
-}
