@@ -1,8 +1,6 @@
-using HahnMovies.Application.Interfaces;
 using HahnMovies.WorkerService;
 using HahnMovies.Infrastructure;
 using HahnMovies.Infrastructure.Data;
-using HahnMovies.Infrastructure.Services;
 using MediatR;
 using Hangfire;
 
@@ -19,10 +17,7 @@ builder.Services.AddHangfire(config =>
     config.UseSqlServerStorage(builder.Configuration.GetConnectionString("Application")));
 builder.Services.AddHangfireServer();
 
-// Register services
-builder.Services.AddScoped<IMovieService, MovieService>();
-builder.Services.AddScoped<ITmdbService, TmdbService>();
-builder.Services.AddScoped<IJobScheduler, HangfireJobScheduler>();
+// Register services;
 
 // Add Hosted Service (Worker)
 builder.Services.AddHostedService<Worker>();
@@ -30,10 +25,5 @@ builder.Services.AddHostedService<Worker>();
 var host = builder.Build();
 
 // Schedule jobs
-using (var scope = host.Services.CreateScope())
-{
-    var jobScheduler = scope.ServiceProvider.GetRequiredService<IJobScheduler>();
-    jobScheduler.ScheduleFullSyncJob();
-}
 
 host.Run();
