@@ -14,10 +14,10 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("HahnMovies.WorkerService")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+5d554044d271384c8f3f9d72f93508fa46e6e638")]
 [assembly: System.Reflection.AssemblyProductAttribute("HahnMovies.WorkerService")]
 [assembly: System.Reflection.AssemblyTitleAttribute("HahnMovies.WorkerService")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]
 
-// Généré par la classe MSBuild WriteCodeFragment.
+// Generated by the MSBuild WriteCodeFragment class.
 
